refactor: presentational components move to correct folder
