+++ app/koi-client/src/component-presentation/StockCard.tsx
