+++ app/koi-client/src/component-presentation/StockLineChart.tsx
